Create CatMouse_a
n = gridsize, s = number of steps allowed to take, t = number of rounds.

A cat and a mouse is put into the grid, cat at coordinates 0,0 and mouse in the middle. Mouse moves in random directions (double steps of cat) and cat moves shortest path towards the mouse. The positions are printed and if cat catches mouse "Catch!" gets printed.