Create CatMouse
A game where a cat chases a mouse. A grid is made and the cat starts in bottom left corner (coordinates 0,0) and the mouse in the middle.
n = gridsize, s = number of steps, t = number of rounds
There are three different versions or methods.
In the first the mouse moves randomly but double the steps of the cat.
The second is identical to the first but there is a kitten who also chases the mouse but takes half the steps as the cat.
And in the third method both cat and mouse takes equal amount of steps but the mouse is smart and will move strategic around the cat until the rounds end and therefore never will be caught.